Add bw-dev create_secrets -command
Command can be used to generate secrets for .env file that are not yet generated. It uses either
python3 secrets module to generate urlsafe token of 50 bytes or openssl to generate hex output of 50
random bytes. 50 bytes is what django 4.2 generates by default, so it should be sane
default for now. 50 random characters is also recommended by owasp (https://cheatsheetseries.owasp.org/cheatsheets/Django_Security_Cheat_Sheet.html#key-management).

Do note that even if the output is alpha-numeric, it is longer than 50 characters, as 50 random bytes are encoded as hex/url_safe values.

Only credential from the list what admin should use is the FLOWER_PASSWORD, but that can be changed
afterwards by admin. Only password that is not trivial to change after initial setup is the
POSTGRES_PASSWORD

Change defaults in settings for SECRET_KEY to be None and complain if SECRET_KEY is not defined.
Comment out secrets from .env.example that should not have default values but unique per deployment

List of .env values that are checked and generated if not enabled in .env:
- SECRET_KEY
- POSTGRES_PASSWORD
- REDIS_ACTIVITY_PASSWORD
- REDIS_BROKER_PASSWORD
- FLOWER_PASSWORD