Add ability to view and terminate user sessions
Adds a dashboard in user preferences to enable users to see all logged-in sessions under their account, and terminate any if they choose to do so.

Also renames 2FA page to "Security Settings" since it seemed reasonable to put these two functionalities together.

Because we use the cache engine for sessions, we can't efficiently check directly for all a user's sessions. Instead we add an object to a new UserSession model the first time they log in to a session. On logout, the actual session is terminated, and we delete the model object. Because sessions can expire, we check all associated sessions for a user's UserSession objects when they load the security page and remove any that no longer have active sessions.

Existing sessions when this is first rolled out will not appear in the dashboard since they won't have a UserSession object. Once they expire, this problem will resolve itself.