Comments out inefficient queries
I decided to leave these queries in place but just not call them so that
they can be re-instated in the future when I or someone else has the
interest in writing them correctly and not repeating my mistake of
making the gnarliest SQL of all time.

I also didn't want to delete them because that would remove the
translations that already exist, and if we do bring them back, we would
lose all the work that translators have put in.