[✨ Features] Add possiblity to create new list from book page #3633
When onto a book page, it is now possible to directly add a new list.
Inside the list dorpdown the last entry is "+ Create new list...". If
the user select this option, a modal is displayed and the user can
directly create the list before confirming. After validation the user
is redirected onto the list page where he can see it's book.